// app-template/openapi/types/eventListItem.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Event Management API
 * API for Event Management System
 * OpenAPI spec version: 1.0.0
 */
import type { EventListItemStatus } from './eventListItemStatus';
import type { EventListItemVisibility } from './eventListItemVisibility';
import type { EventListItemRegistrationStatus } from './eventListItemRegistrationStatus';

export interface EventListItem {
  id: string;
  name: string;
  slug: string;
  startDate: Date;
  endDate: Date;
  /** @nullable */
  venueName?: string | null;
  status: EventListItemStatus;
  visibility: EventListItemVisibility;
  registrationStatus: EventListItemRegistrationStatus;
  /** @nullable */
  capacity?: number | null;
  registeredCount?: number;
  checkedInCount?: number;
  createdAt: Date;
  updatedAt: Date;
}